backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.10
django-cleanup==8.1.0
django-debug-toolbar==4.3.0
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
dotenv==0.9.9
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.1.1
flake8-broken-line==1.0.0
flake8-docstrings==1.7.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
fonttools==4.57.0
fpdf2==2.7.8
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pep8==1.7.1
pep8-naming==0.14.1
pillow==11.1.0
pluggy==1.5.0
psutil==6.1.1
psycopg2-binary==2.9.10
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.10.1
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.8.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uharfbuzz==0.39.1
urllib3==2.3.0
